backend/mole_model/requirements.txt
albucore==0.0.23
albumentations==2.0.5
alembic==1.15.2
annotated-types==0.7.0
bleach==6.2.0
catboost==1.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
graphviz==0.20.3
greenlet==3.1.1
h5py==3.13.0
huggingface-hub==0.30.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.6
joblib==1.4.2
kaggle==1.7.4.2
kagglehub==0.3.11
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.33.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
polars==1.26.0
protobuf==6.30.2
pydantic==2.11.2
pydantic_core==2.33.1
pydealer==1.4.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.40
stringzilla==3.12.3
sympy==1.13.1
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1
xgboost==3.0.0
